pyproject.toml
[tool.poetry]
name = "repl_nix_Fuzionimo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
replit = "^3.2.5"
requests = "^2.28.2"
wget = "^3.2"
discord = "^2.2.2"
Pillow = "^9.4.0"
discord-py = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
